+++ Changelog.md
@@ -1,6 +1,9 @@
 # Change Log
 
-## bpaf [0.9.1] - 2023-07-05, bpaf_derive [0.5.1]
+## bpaf [0.9.2], bpaf_derive [0.5.2] - Unreleased
+- with `docgen` feature you can render documentation as markdown
+
+## bpaf [0.9.1], bpaf_derive [0.5.1] - 2023-07-05
 - add a way to print usage when called with no argument_os
 - since 0.9.0 bpaf splits help messages into "full" and "partial", displaying
   full only when `--help` flag is passed twice or when rendering the documentation,